- changes to api endpoint routes - finished gallery api endpoints

+++ backend/api/admin_settings.py
@@ -23,7 +23,7 @@ post_account_model = admin_settings_ns.model('Account', {
 
 # Define a resource for the '/admin_settings' endpoint
 # Used to get all accounts on admin settings page
-@admin_settings_ns.route('/admin_settings', methods=['GET', 'POST', 'DELETE']) 
+@admin_settings_ns.route('/', methods=['GET', 'POST', 'DELETE']) 
 class AdminSettingsResource(Resource):
     @admin_settings_ns.marshal_list_with(get_account_model)
     #@jwt_required() # Protect this endpoint with JWT
@@ -44,6 +44,7 @@ class AdminSettingsResource(Resource):
 
         if db_user is not None:
             return {'message': f'Username {username} already exists'}, 400
+            
 
         # Create a new account if the username does not exist
         new_account = Account(name=data.get('name'), username=data.get('username'), password=generate_password_hash(data.get('password')), role=data.get('role'))
@@ -54,7 +55,7 @@ class AdminSettingsResource(Resource):
     
 # Define a resource for the '/admin_settings/<string:username>' endpoint
 # Used to get a specific account on admin settings page when entered in the search bar
-@admin_settings_ns.route('/admin_settings/<string:username>', methods=['GET'])
+@admin_settings_ns.route('/<string:username>', methods=['GET'])
 class AdminSettingResource(Resource):
     @admin_settings_ns.marshal_with(get_account_model)
     #@jwt_required() # Protect this endpoint with JWT
@@ -66,11 +67,15 @@ class AdminSettingResource(Resource):
 
 # Define a resource for the '/admin_settings/<int:account_id>' endpoint
 # Used to delete a specific account on admin settings page
-@admin_settings_ns.route('/admin_settings/<int:account_id>', methods=['DELETE'])
+@admin_settings_ns.route('/<int:account_id>', methods=['DELETE'])
 class AdminSettingResource(Resource):
     #@jwt_required() # Protect this endpoint with JWT
     def delete(self, account_id):
         account = Account.query.get(account_id)
+
+        # Check if the account exists
+        if account is None:
+            return {'message': 'Account not found'}, 404
         
         account.delete()
